fix: close z21 qa findings
