+++ crates/atm/Cargo.toml
@@ -18,6 +18,9 @@ exclude = ["src/bin/atm_post_send_hook_fixture.rs"]
 name = "atm"
 path = "src/main.rs"
 
+[features]
+fault-injection = ["sc-observability/fault-injection"]
+
 [dependencies]
 atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "1.2.0" }
 atm-daemon-bootstrap = { path = "../atm-daemon-bootstrap", version = "1.2.0" }
