+++ scripts/test_atm_nudge.py
@@ -25,6 +25,9 @@ ERR_INVALID_STRUCTURE = _MOD.ERR_INVALID_STRUCTURE
 ERR_NOT_FOUND = _MOD.ERR_NOT_FOUND
 ERR_PARSE_ERROR = _MOD.ERR_PARSE_ERROR
 CODEX_DEFAULT_PANE = _MOD.CODEX_DEFAULT_PANE
+TEST_TEAM = "atm-dev"
+TEST_AGENT = "arch-ctm"
+TEST_TEAM_LEAD = "team-lead"
 
 
 def _parse_json(text: str) -> dict:
@@ -39,7 +42,7 @@ def _run_with_mocked_lookups(
     toml: PaneLookup,
     cfg: PaneLookup,
     *,
-    team: str = "atm-dev",
+    team: str = TEST_TEAM,
 ) -> tuple[int, dict, dict, MagicMock]:
     stderr_buf = io.StringIO()
     stdout_buf = io.StringIO()
@@ -66,16 +69,16 @@ class TestNudgePane(unittest.TestCase):
 
     def test_valid_inputs_accepted(self):
         with patch("subprocess.run") as mock_run, patch.object(_MOD, "log"):
-            _MOD.nudge_pane("%1", "arch-ctm", "<atm/>")
+            _MOD.nudge_pane("%1", TEST_AGENT, "<atm/>")
         self.assertEqual(mock_run.call_count, 2)
 
     def test_empty_pane_raises(self):
         with self.assertRaises(ValueError):
-            self._call("", "arch-ctm", "<atm/>")
+            self._call("", TEST_AGENT, "<atm/>")
 
     def test_whitespace_pane_raises(self):
         with self.assertRaises(ValueError):
-            self._call("   ", "arch-ctm", "<atm/>")
+            self._call("   ", TEST_AGENT, "<atm/>")
 
     def test_empty_recipient_raises(self):
         with self.assertRaises(ValueError):
@@ -83,11 +86,11 @@ class TestNudgePane(unittest.TestCase):
 
     def test_empty_message_raises(self):
         with self.assertRaises(ValueError):
-            self._call("%1", "arch-ctm", "")
+            self._call("%1", TEST_AGENT, "")
 
     def test_non_string_pane_raises(self):
         with self.assertRaises(ValueError):
-            self._call(None, "arch-ctm", "<atm/>")
+            self._call(None, TEST_AGENT, "<atm/>")
 
     def test_tmux_calls_order(self):
         with patch("subprocess.run") as mock_run, patch.object(_MOD, "log"):
@@ -160,14 +163,14 @@ class TestReadPaneFromToml(unittest.TestCase):
 
     def test_reads_team_specific_match(self):
         def run(path: Path):
-            result = _MOD.read_pane_from_toml("quality-mgr", "atm-dev")
+            result = _MOD.read_pane_from_toml("quality-mgr", TEST_TEAM)
             self.assertEqual(result.pane_id, "%2")
             self.assertEqual(Path(result.source_path), path.resolve())
 
         self._with_project(
-            """
+            f"""
 [atm]
-default_team = "atm-dev"
+default_team = "{TEST_TEAM}"
 
 [rmux]
 
@@ -176,31 +179,31 @@ name = "agents"
 [[rmux.windows.panes]]
 name = "quality-mgr"
 tmux_pane_id = "%2"
-env = { ATM_TEAM = "atm-dev" }
+env = {{ ATM_TEAM = "{TEST_TEAM}" }}
 [[rmux.windows.panes]]
 name = "quality-mgr"
 tmux_pane_id = "%9"
-env = { ATM_TEAM = "schook" }
+env = {{ ATM_TEAM = "schook" }}
 """,
             run,
         )
 
     def test_falls_back_to_single_unscoped_match(self):
         def run(_path: Path):
-            result = _MOD.read_pane_from_toml("arch-ctm", "atm-dev")
+            result = _MOD.read_pane_from_toml(TEST_AGENT, TEST_TEAM)
             self.assertEqual(result.pane_id, "%1")
 
         self._with_project(
-            """
+            f"""
 [atm]
-default_team = "atm-dev"
+default_team = "{TEST_TEAM}"
 
 [rmux]
 
 [[rmux.windows]]
 name = "agents"
 [[rmux.windows.panes]]
-name = "arch-ctm"
+name = "{TEST_AGENT}"
 tmux_pane_id = "%1"
 """,
             run,
@@ -208,15 +211,15 @@ tmux_pane_id = "%1"
 
     def test_reports_ambiguous_same_team_match(self):
         def run(_path: Path):
-            result = _MOD.read_pane_from_toml("quality-mgr", "atm-dev")
+            result = _MOD.read_pane_from_toml("quality-mgr", TEST_TEAM)
             self.assertEqual(result.error_code, ERR_AMBIGUOUS)
             self.assertIn("%2", result.error_msg)
             self.assertIn("%7", result.error_msg)
 
         self._with_project(
-            """
+            f"""
 [atm]
-default_team = "atm-dev"
+default_team = "{TEST_TEAM}"
 
 [rmux]
 
@@ -225,18 +228,18 @@ name = "agents"
 [[rmux.windows.panes]]
 name = "quality-mgr"
 tmux_pane_id = "%2"
-env = { ATM_TEAM = "atm-dev" }
+env = {{ ATM_TEAM = "{TEST_TEAM}" }}
 [[rmux.windows.panes]]
 name = "quality-mgr"
 tmux_pane_id = "%7"
-env = { ATM_TEAM = "atm-dev" }
+env = {{ ATM_TEAM = "{TEST_TEAM}" }}
 """,
             run,
         )
 
     def test_reports_parse_error(self):
         def run(path: Path):
-            result = _MOD.read_pane_from_toml("arch-ctm", "atm-dev")
+            result = _MOD.read_pane_from_toml(TEST_AGENT, TEST_TEAM)
             self.assertEqual(result.error_code, ERR_PARSE_ERROR)
             self.assertIn(str(path), result.error_msg)
 
@@ -247,7 +250,7 @@ env = { ATM_TEAM = "atm-dev" }
             root = Path(tmp)
             with patch.dict(os.environ, {"CLAUDE_PROJECT_DIR": str(root), "PWD": str(root)}, clear=False):
                 with patch("os.getcwd", return_value=str(root)):
-                    result = _MOD.read_pane_from_toml("arch-ctm", "atm-dev")
+                    result = _MOD.read_pane_from_toml(TEST_AGENT, TEST_TEAM)
         self.assertEqual(result.error_code, ERR_FILE_MISSING)
 
 
@@ -255,11 +258,11 @@ class TestReadPaneFromConfig(unittest.TestCase):
     def test_reports_invalid_members_structure(self):
         with tempfile.TemporaryDirectory() as tmp:
             home = Path(tmp)
-            cfg = home / ".claude" / "teams" / "atm-dev"
+            cfg = home / ".claude" / "teams" / TEST_TEAM
             cfg.mkdir(parents=True)
             (cfg / "config.json").write_text('{"members": {}}', encoding="utf-8")
             with patch.object(Path, "home", return_value=home):
-                result = _MOD.read_pane_from_config("arch-ctm", "atm-dev")
+                result = _MOD.read_pane_from_config(TEST_AGENT, TEST_TEAM)
         self.assertEqual(result.error_code, ERR_INVALID_STRUCTURE)
 
 
@@ -284,42 +287,42 @@ class TestOverrideMode(unittest.TestCase):
             patch.object(_MOD, "nudge_pane") as mock_nudge,
             patch.object(_MOD, "read_pane_from_toml") as mock_toml,
             patch.object(_MOD, "read_pane_from_config") as mock_cfg,
-            patch.object(_MOD, "resolve_team", return_value="atm-dev"),
+            patch.object(_MOD, "resolve_team", return_value=TEST_TEAM),
         ):
-            rc = _MOD.main(["atm-nudge.py", "--pane", "%1", "arch-ctm", "<atm/>"])
+            rc = _MOD.main(["atm-nudge.py", "--pane", "%1", TEST_AGENT, "<atm/>"])
         self.assertEqual(rc, 0)
-        mock_nudge.assert_called_once_with("%1", "arch-ctm", "<atm/>")
+        mock_nudge.assert_called_once_with("%1", TEST_AGENT, "<atm/>")
         mock_toml.assert_not_called()
         mock_cfg.assert_not_called()
 
     def test_override_without_message_builds_default(self):
         with (
             patch.object(_MOD, "nudge_pane") as mock_nudge,
-            patch.object(_MOD, "resolve_team", return_value="atm-dev"),
+            patch.object(_MOD, "resolve_team", return_value=TEST_TEAM),
             patch.object(_MOD, "read_post_send_payload", return_value={}),
             patch.object(_MOD, "read_pane_from_toml"),
             patch.object(_MOD, "read_pane_from_config"),
         ):
-            rc = _MOD.main(["atm-nudge.py", "--pane", "%1", "arch-ctm"])
+            rc = _MOD.main(["atm-nudge.py", "--pane", "%1", TEST_AGENT])
         self.assertEqual(rc, 0)
         _, recipient, message = mock_nudge.call_args[0]
-        self.assertEqual(recipient, "arch-ctm")
-        self.assertIn("read atm --team atm-dev", message)
+        self.assertEqual(recipient, TEST_AGENT)
+        self.assertIn(f"read atm --team {TEST_TEAM}", message)
 
 
 class TestBuildMessage(unittest.TestCase):
     def test_default_send_message_requests_assigned_task_execution(self):
-        message = _MOD.build_message("atm-dev", {})
-        self.assertIn("read atm --team atm-dev", message)
+        message = _MOD.build_message(TEST_TEAM, {})
+        self.assertIn(f"read atm --team {TEST_TEAM}", message)
         self.assertIn("execute the assigned task", message)
         self.assertIn('busy="after-current-task"', message)
 
     def test_ack_message_requests_immediate_work_with_message_context(self):
         message = _MOD.build_message(
-            "atm-dev",
+            TEST_TEAM,
             {"is_ack": True, "message_id": "01JACKTEST00000000000000000"},
         )
-        self.assertIn("read atm --team atm-dev", message)
+        self.assertIn(f"read atm --team {TEST_TEAM}", message)
         self.assertIn("message 01JACKTEST00000000000000000 acknowledged", message)
         self.assertIn("complete associated work immediately", message)
         self.assertIn(
@@ -332,23 +335,23 @@ class TestBuildMessage(unittest.TestCase):
 class TestMainBehavior(unittest.TestCase):
     def test_matching_panes_nudges_without_warning(self):
         rc, stderr_json, stdout_json, mock_nudge = _run_with_mocked_lookups(
-            ["arch-ctm"],
+            [TEST_AGENT],
             PaneLookup("%1", None, None, "/repo/.atm.toml"),
             PaneLookup("%1", None, None, "/home/config.json"),
         )
         self.assertEqual(rc, 0)
-        mock_nudge.assert_called_once_with("%1", "arch-ctm", unittest.mock.ANY)
+        mock_nudge.assert_called_once_with("%1", TEST_AGENT, unittest.mock.ANY)
         self.assertEqual(stderr_json, {})
         self.assertEqual(stdout_json, {})
 
     def test_config_mismatch_still_nudges_and_warns(self):
         rc, stderr_json, stdout_json, mock_nudge = _run_with_mocked_lookups(
-            ["arch-ctm"],
+            [TEST_AGENT],
             PaneLookup("%1", None, None, "/repo/.atm.toml"),
             PaneLookup("%9", None, None, "/home/config.json"),
         )
         self.assertEqual(rc, 0)
-        mock_nudge.assert_called_once_with("%1", "arch-ctm", unittest.mock.ANY)
+        mock_nudge.assert_called_once_with("%1", TEST_AGENT, unittest.mock.ANY)
         self.assertEqual(stderr_json["status"], "warning")
         self.assertIn("pane %1", " ".join(stderr_json["call_to_action"]))
         self.assertIn("config.json", " ".join(stderr_json["call_to_action"]))
@@ -387,7 +390,7 @@ class TestMainBehavior(unittest.TestCase):
 
     def test_neither_source_found_uses_default_pane(self):
         rc, stderr_json, stdout_json, mock_nudge = _run_with_mocked_lookups(
-            ["arch-ctm"],
+            [TEST_AGENT],
             PaneLookup(None, ERR_NOT_FOUND, "missing recipient", "/repo/.atm.toml"),
             PaneLookup(None, ERR_NOT_FOUND, "missing member", "/home/config.json"),
         )
@@ -399,14 +402,14 @@ class TestMainBehavior(unittest.TestCase):
 
     def test_error_payload_includes_input_and_resolution_context(self):
         rc, stderr_json, _, _ = _run_with_mocked_lookups(
-            ["arch-ctm"],
+            [TEST_AGENT],
             PaneLookup(None, ERR_FILE_MISSING, "missing", None),
             PaneLookup(None, ERR_FILE_MISSING, "missing", "/home/config.json"),
         )
         self.assertEqual(rc, 1)
         self.assertIn("input", stderr_json)
         self.assertIn("pane_resolution", stderr_json)
-        self.assertEqual(stderr_json["input"]["recipient"], "arch-ctm")
+        self.assertEqual(stderr_json["input"]["recipient"], TEST_AGENT)
         self.assertEqual(stderr_json["pane_resolution"]["authoritative_source"], ".atm.toml")
